+++ modules/astparser/src/ast/declarations/ASTNamespaceDeclaration.h
@@ -59,7 +59,7 @@ public:
     }
 
     virtual ~ASTNamespaceDeclaration() {
-        for (unsigned int i = 0, count = declarations.size(); i < count; i++)
+        for (size_t i = 0, count = declarations.size(); i < count; i++)
             delete declarations[i];
         declarations.clear();
     }
